re-introduce splash page
